refactor: github chat.
